astro/requirements.txt
apache-airflow-providers-amazon
apache-airflow-providers-snowflake
apache-airflow
Office365-REST-Python-Client
office365
webdriver-manager
selenium
numpy
openpyxl
pandas
pytest
requests